feat: Review에 rating 타입 Integer로 변경

+++ src/main/java/com/sprint/deokhugam/domain/review/service/ReviewServiceImpl.java
@@ -114,7 +114,7 @@ public class ReviewServiceImpl implements ReviewService {
         validateDuplicateReview(bookId, userId);
 
         String content = request.content();
-        double rating = request.rating();
+        Integer rating = request.rating();
 
         Review review = new Review(rating, content, book, user);
         Review savedReview = reviewRepository.save(review);
